+++ ofdrw-layout/src/test/java/org/ofdrw/layout/OFDDocTest.java
@@ -39,6 +39,22 @@ import static org.junit.jupiter.api.Assertions.*;
 class OFDDocTest {
 
     /**
+     * 字体宽度溢出可用最大宽度测试
+     */
+    @Test
+    void fontSizeOverflow() throws IOException {
+        Path outP = Paths.get("target/FontSizeOverflow.ofd");
+
+        try (OFDDoc ofdDoc = new OFDDoc(outP)) {
+            Paragraph p = new Paragraph(10d, 20d).setFontSize(15d);
+            p.add("l我l");
+            ofdDoc.add(p);
+            // Expect: 只显示 "l"
+        }
+        System.out.println("生成文档位置：" + outP.toAbsolutePath().toString());
+    }
+
+    /**
      * 向文件中加入附件文件
      *
      * @throws IOException
@@ -626,6 +642,36 @@ class OFDDocTest {
             ofdDoc.add(p)
                     .add(canvas)
                     .add(p2);
+        }
+        System.out.println("生成文档位置：" + path.toAbsolutePath());
+    }
+
+    @Test
+    public void virtualPagePositioning() throws IOException {
+        Path path = Paths.get("target/virtualPagePositioning.ofd").toAbsolutePath();
+        try (OFDDoc ofdDoc = new OFDDoc(path)) {
+
+            VirtualPage vPage = new VirtualPage(ofdDoc.getPageLayout());
+
+            Canvas canvas = new Canvas(200d, 200d);
+            canvas.setPosition(Position.Absolute)
+                    .setX(0d)
+                    .setY(0d)
+                    .setBorder(1d);
+            canvas.setDrawer(ctx -> {
+                ctx.beginPath();
+                ctx.arc(10, 10, 5, 0, 360);
+                ctx.stroke();
+            });
+
+            vPage.add(canvas);
+            ofdDoc.addVPage(vPage);
+
+            Paragraph p = new Paragraph("是不是很好看是不是很好看是不是很好看是不是很好看");
+            p.setPosition(Position.Absolute)
+                    .setWidth(30d)
+                    .setX(50d).setY(100d);
+            vPage.add(p);
         }
         System.out.println("生成文档位置：" + path.toAbsolutePath());
     }